Wonderful things.
Added hasInput and hasOutput relations in relation.h to characterize
model behaviour.

Moved all the relation related to models under hasSign class as they are
a semiotic relation type.

Added the removeRelation method in Thing.

Created the LatexExpression class and used it as part of default
constructor for Model class and all its chidrens.

Minor changes in knowledgegenerators.h on ismodelfor method.

Added hasInput, hasOutput and mathematical expressions in constructors
for surfacetensionmodel.h and saturationpressuremodel.h.