New Models and features.
Added the MomentModelPratsinis and ClassicalNucleationTheory software
models.

Ane example of usage has been added in main.cpp.

At the moment, objects which has the same ontological class but different
physical meaning ( the Temperature of a gas and the Melting point of a
substance ) have been distinguished using a label. This label allows to
to use the getLabeledRelatedObjects method which return a pointers list
of objects which have the desired class and label.

In order to reduce the code necessary to get the data stored in a
Quantity, the methods onData() and onUnit() were defined. These methods
directly return the pointer to the data/unit.

All the models, methods and examples now make large use of pointers to
access and modify data. In this way the ontological objects are modified
run-time without need to update them manually.

In future releases a method to copy a class and its relations will be
implemented.